requirements.txt
kivy_garden.graph
pymodbustcp
kivy[full]
